refactor(cli): move command classes out of ruyi.cli
