+++ ruyi/ruyipkg/news_cli.py
@@ -3,10 +3,11 @@ import argparse
 from rich import box
 from rich.table import Table
 
+from .. import is_porcelain, log
+from ..cli.cmd import RootCommand
 from ..config import GlobalConfig
 from ..utils.markdown import RuyiStyledMarkdown
 from ..utils.porcelain import PorcelainOutput
-from .. import is_porcelain, log
 from .news import NewsItem, NewsItemContent, NewsItemStore
 from .repo import MetadataRepo
 
@@ -34,53 +35,98 @@ def print_news_item_titles(
     log.stdout(tbl)
 
 
-def cli_news_list(config: GlobalConfig, args: argparse.Namespace) -> int:
-    only_unread = args.new
+class NewsCommand(
+    RootCommand,
+    cmd="news",
+    has_subcommands=True,
+    help="List and read news items from configured repository",
+):
+    pass
 
-    mr = MetadataRepo(config)
-    store = mr.news_store()
-    newsitems = store.list(only_unread)
 
-    if is_porcelain():
-        with PorcelainOutput() as po:
-            for ni in newsitems:
-                po.emit(ni.to_porcelain())
+class NewsListCommand(
+    NewsCommand,
+    cmd="list",
+    help="List news items",
+):
+    @classmethod
+    def configure_args(cls, p: argparse.ArgumentParser) -> None:
+        p.add_argument(
+            "--new",
+            action="store_true",
+            help="List unread news items only",
+        )
+
+    @classmethod
+    def main(cls, cfg: GlobalConfig, args: argparse.Namespace) -> int:
+        only_unread = args.new
+
+        mr = MetadataRepo(cfg)
+        store = mr.news_store()
+        newsitems = store.list(only_unread)
+
+        if is_porcelain():
+            with PorcelainOutput() as po:
+                for ni in newsitems:
+                    po.emit(ni.to_porcelain())
+            return 0
+
+        log.stdout("[bold green]News items:[/bold green]\n")
+        if not newsitems:
+            log.stdout("  (no unread item)" if only_unread else "  (no item)")
+            return 0
+
+        print_news_item_titles(newsitems, cfg.lang_code)
+
         return 0
 
-    log.stdout("[bold green]News items:[/bold green]\n")
-    if not newsitems:
-        log.stdout("  (no unread item)" if only_unread else "  (no item)")
+
+class NewsReadCommand(
+    NewsCommand,
+    cmd="read",
+    help="Read news items",
+    description="Outputs news item(s) to the console and mark as already read. Defaults to reading all unread items if no item is specified.",
+):
+    @classmethod
+    def configure_args(cls, p: argparse.ArgumentParser) -> None:
+        p.add_argument(
+            "--quiet",
+            "-q",
+            action="store_true",
+            help="Do not output anything and only mark as read",
+        )
+        p.add_argument(
+            "item",
+            type=str,
+            nargs="*",
+            help="Ordinal or ID of the news item(s) to read",
+        )
+
+    @classmethod
+    def main(cls, cfg: GlobalConfig, args: argparse.Namespace) -> int:
+        quiet = args.quiet
+        items_strs = args.item
+
+        mr = MetadataRepo(cfg)
+        store = mr.news_store()
+
+        # filter out requested news items
+        items = filter_news_items_by_specs(store, items_strs)
+        if items is None:
+            return 1
+
+        # render the items
+        if not quiet:
+            if items:
+                for ni in items:
+                    print_news(ni.get_content_for_lang(cfg.lang_code))
+            else:
+                log.stdout("No news to display.")
+
+        # record read statuses
+        store.mark_as_read(*(ni.id for ni in items))
+
         return 0
-
-    print_news_item_titles(newsitems, config.lang_code)
-
-    return 0
-
-
-def cli_news_read(config: GlobalConfig, args: argparse.Namespace) -> int:
-    quiet = args.quiet
-    items_strs = args.item
-
-    mr = MetadataRepo(config)
-    store = mr.news_store()
-
-    # filter out requested news items
-    items = filter_news_items_by_specs(store, items_strs)
-    if items is None:
-        return 1
-
-    # render the items
-    if not quiet:
-        if items:
-            for ni in items:
-                print_news(ni.get_content_for_lang(config.lang_code))
-        else:
-            log.stdout("No news to display.")
-
-    # record read statuses
-    store.mark_as_read(*(ni.id for ni in items))
-
-    return 0
 
 
 def filter_news_items_by_specs(
